Added Retina display support and improved lander movement thru use of floats for X/Y screen position
Increased the display update rate so we are less likely to move multiple pixels in a display update (causing jerky movement).  One side effect is that we need to decouple the dust/thrust updates from the display update since we lose information if these are done too frequently.

+++ Moonlander/Classes/Dust.m
@@ -57,7 +57,7 @@ const float DustViewHeight = 64;
             }
             
             //(DUSTP1)  Thrust angle determines dust direction
-            short deltaY = self.delegate.SHOWY - self.delegate.AVERT;
+            short deltaY = (short)self.delegate.SHOWY - self.delegate.AVERT;
             float sinDeltaY = deltaY * sinAngle;
             float cosAngle = cos(angle);
             float tanDeltaY = sinDeltaY / cosAngle;
